+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+1.0.1 / 2026-02-13
+==================
+- Added check on event keyCode as well as key to prevent empty or unidentified codes from being detected.
+- Added options interface for setting preventDefault on different instances of the barcode detector.
+- Added `index.d.ts` file for containing TypeScript interfaces, and moved them from `index.ts`.
+
 1.0.0 / 2024-11-29
 ==================
 - Initial release